requirements.txt
altgraph==0.17.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.2.0
auto-py-to-exe==2.13.0
az-iranian-bank-gateways==1.6.15
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
bottle==0.12.19
bottle-websocket==0.2.9
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-extensions==3.1.5
django-mathfilters==1.0.0
django-render-partial==0.4
djangorestframework==3.13.1
Eel==0.12.4
entrypoints==0.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
hiredis==2.0.0
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg==3.0.7
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycparser==2.21
pycryptodome==3.12.0
pydot==1.4.2
Pygments==2.10.0
pygraphviz==1.7
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
redis==4.1.0
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
whichcraft==0.6.1
widgetsnbextension==3.5.2
wrapt==1.13.3
zeep==4.1.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
